praskn_InterviewBot/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
fastui==0.5.2
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.6
mistralai==0.1.8
orjson==3.10.0
pydantic==2.6.4
pydantic_core==2.16.3
python-decouple==3.8
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
uvicorn==0.29.0